remove client dependency; add RefCache.size command
Search functionality is removed from RefCache because it was way simpler
than what we need in production in the collectionspace-mapper. It is
easier and clearer to search using collectionspace-client directly
(and then add result to RefCache) in applications that need both direct
interaction with a CS instance and caching.

This removes the need to have collectionspace-client as a dependency, which
decouples RefCache from Client, making development on each simpler.

Added RefCache.size (and #{backend}.size) methods since this is useful to have in
applications using RefCache.