feat: Add optional `value_type` param to caching commands
This allows for distinct CSID, refname, URI, or any other values for a
record to be stored in the same cache by including the value type in
the hash key if it is provided.

If not provided, `value_type` param defaults to nil, and is not
included in cached key. This supports current functionality as used by
collectionspace-mapper by the CSV Importer and
collectionspace_migration_tools, both of which use two separate caches
for CSIDs and refnames